test broker and move message
